requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
pywebview>=4.0.0
PyYAML>=6.0
websockets>=12.0
firebase-admin>=6.0.0
pynput>=1.7.6
plyer>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
